chatbot/requirements.txt
fastapi
uvicorn
pandas
geopandas
xgboost
scikit-learn
google-generativeai
python-dotenv
shapely
